Moved database related files to data folder
