+++ src/main/java/com/honiism/discord/lemi/Lemi.java
@@ -31,7 +31,7 @@ import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.honiism.discord.lemi.commands.slash.currency.objects.items.Items;
 import com.honiism.discord.lemi.commands.slash.handler.ISlashCmd;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmdManager;
-import com.honiism.discord.lemi.database.managers.LemiDbManager;
+import com.honiism.discord.lemi.data.database.managers.LemiDbManager;
 import com.honiism.discord.lemi.listeners.BaseListener;
 import com.honiism.discord.lemi.listeners.CustomEmbedListener;
 import com.honiism.discord.lemi.listeners.GuildListener;
