+++ src/main/java/com/honiism/discord/lemi/commands/slash/staff/admins/Embed.java
@@ -24,7 +24,7 @@ import java.util.HashMap;
 import com.honiism.discord.lemi.commands.handler.CommandCategory;
 import com.honiism.discord.lemi.commands.handler.UserCategory;
 import com.honiism.discord.lemi.commands.slash.handler.SlashCmd;
-import com.honiism.discord.lemi.database.managers.LemiDbEmbedManager;
+import com.honiism.discord.lemi.data.database.managers.LemiDbEmbedManager;
 import com.honiism.discord.lemi.utils.customEmbeds.EmbedTools;
 import com.honiism.discord.lemi.utils.misc.Tools;
